Stubbed out RPC server, now client and server can connect. Pid mapping does not allow file creation at this time.

+++ kernel/src/arch/x86_64/syscall.rs
@@ -382,7 +382,7 @@ fn handle_fileio(
                 let filename = userptr_to_str(pathname)?;
                 let mut client = kcb.arch.rpc_client.lock();
                 return match client.as_mut().unwrap().fio_open(
-                    pid as u64,
+                    pid,
                     filename.as_bytes(),
                     flags,
                     modes,
@@ -410,12 +410,11 @@ fn handle_fileio(
 
                 if op == FileOperation::Read {
                     let mut userslice = super::process::UserSlice::new(buffer, len as usize);
-                    return match client.as_mut().unwrap().fio_read(
-                        pid as u64,
-                        fd,
-                        len,
-                        &mut userslice,
-                    ) {
+                    return match client
+                        .as_mut()
+                        .unwrap()
+                        .fio_read(pid, fd, len, &mut userslice)
+                    {
                         Ok(a) => Ok(a),
                         Err(err) => Err(err.into()),
                     };
@@ -423,11 +422,11 @@ fn handle_fileio(
                     // write operation
                     let kernslice = crate::process::KernSlice::new(buffer, len as usize);
                     let buff_ptr = kernslice.buffer.clone();
-                    return match client.as_mut().unwrap().fio_write(
-                        pid as u64,
-                        fd,
-                        buff_ptr.to_vec(),
-                    ) {
+                    return match client
+                        .as_mut()
+                        .unwrap()
+                        .fio_write(pid, fd, buff_ptr.to_vec())
+                    {
                         Ok(a) => Ok(a),
                         Err(err) => Err(err.into()),
                     };
@@ -454,7 +453,7 @@ fn handle_fileio(
                 if op == FileOperation::ReadAt {
                     let mut userslice = super::process::UserSlice::new(buffer, len as usize);
                     return match client.as_mut().unwrap().fio_readat(
-                        pid as u64,
+                        pid,
                         fd,
                         len,
                         offset,
@@ -468,7 +467,7 @@ fn handle_fileio(
                     let kernslice = crate::process::KernSlice::new(buffer, len as usize);
                     let buff_ptr = kernslice.buffer.clone();
                     return match client.as_mut().unwrap().fio_writeat(
-                        pid as u64,
+                        pid,
                         fd,
                         offset,
                         buff_ptr.to_vec(),
@@ -498,7 +497,6 @@ fn handle_fileio(
 
             #[cfg(not(feature = "exokernel"))]
             {
-                #[cfg(not(feature = "exokernel"))]
                 return cnrfs::MlnrKernelNode::unmap_fd(pid, fd);
             }
         }
@@ -517,7 +515,7 @@ fn handle_fileio(
                 return match client
                     .as_mut()
                     .unwrap()
-                    .fio_getinfo(pid as u64, filename.as_bytes())
+                    .fio_getinfo(pid, filename.as_bytes())
                 {
                     Ok((ftype, fsize)) => {
                         let user_ptr = UserPtr::new(&mut VAddr::from(info_ptr));
@@ -547,7 +545,7 @@ fn handle_fileio(
                 return match client
                     .as_mut()
                     .unwrap()
-                    .fio_delete(pid as u64, filename.as_bytes())
+                    .fio_delete(pid, filename.as_bytes())
                 {
                     Ok(a) => Ok(a),
                     Err(err) => Err(err.into()),
@@ -595,7 +593,7 @@ fn handle_fileio(
                 let newname = userptr_to_str(newname)?;
                 let mut client = kcb.arch.rpc_client.lock();
                 return match client.as_mut().unwrap().fio_rename(
-                    pid as u64,
+                    pid,
                     oldname.as_bytes(),
                     newname.as_bytes(),
                 ) {
@@ -619,11 +617,11 @@ fn handle_fileio(
                 // TODO
                 let pathname = userptr_to_str(pathname)?;
                 let mut client = kcb.arch.rpc_client.lock();
-                return match client.as_mut().unwrap().fio_mkdir(
-                    pid as u64,
-                    pathname.as_bytes(),
-                    modes,
-                ) {
+                return match client
+                    .as_mut()
+                    .unwrap()
+                    .fio_mkdir(pid, pathname.as_bytes(), modes)
+                {
                     Ok(a) => Ok(a),
                     Err(err) => Err(err.into()),
                 };
